// src/api/bindings/BillingSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ReceiptPaper } from "./ReceiptPaper";

export type BillingSettings = { currency: string, receiptPrefix: string, receiptDigits: number, paymentMethods: Array<string>, receiptFooter: string, receiptPaper: ReceiptPaper, 
/**
 * Open the print dialog automatically after saving a payment.
 */
autoPrint: boolean, };